rename: index.ts -> types.ts
