+++ src/features/shared/icons/bookmark.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from ".";
+import type { IconProps } from "./types";
 
 export default function BookMark({
   title,
