+++ src/features/shared/icons/file.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from ".";
+import type { IconProps } from "./types";
 
 const FileEditIcon = ({
   title,
