+++ src/features/shared/icons/password.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from ".";
+import type { IconProps } from "./types";
 
 export default function PasswordIcon({
   title,
