+++ src/features/shared/icons/report.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from ".";
+import type { IconProps } from "./types";
 
 export default function Report({
   title,
